refactor model classes into separate module
